Implement external data injestion from Hiera
Turns our Hiera can use variables in returning values, but always casts
them to string. So if you have a scope variable with Arbiter data, and
you try to use that by setting a Hiera key to resolve to that value,
it'll cast it to string first, and you have to re-cast it back to a Hash
in change_risk to make it usable.