Fix block-form evaluation order bug
This one was tricky to identify because the behavior didn't seem to be
consistent, but... the eval_noop() function needs to be called on a
scope *before* the lambda is evaluated in that scope, in order for the
noop() to propogate correctly.

Seems obvious once you figure it out. Something about it though made the
behavior really inconsistent, and so the problem difficult to track
down.